(PC-16709) refactor(cookies): create helpers folder
