+++ src/features/auth/AuthContext.tsx
@@ -7,7 +7,7 @@ import { refreshAccessToken } from 'api/apiHelpers'
 import { SigninResponse } from 'api/gen'
 import { useResetContexts } from 'features/auth/useResetContexts'
 import { usePostCookiesConsent } from 'features/cookies/api/usePostCookiesConsent'
-import { useCookies } from 'features/cookies/useCookies'
+import { useCookies } from 'features/cookies/helpers/useCookies'
 import { useAppStateChange } from 'libs/appState'
 import { analytics, LoginRoutineMethod } from 'libs/firebase/analytics'
 import { getAccessTokenStatus, getUserIdFromAccesstoken } from 'libs/jwt'
