+++ src/features/cookies/api/usePostCookiesConsent.ts
@@ -1,7 +1,7 @@
 import { useMutation } from 'react-query'
 
 import { api } from 'api/api'
-import { getCookiesChoice } from 'features/cookies/useCookies'
+import { getCookiesChoice } from 'features/cookies/helpers/useCookies'
 
 export function usePostCookiesConsent() {
   return useMutation(async () => {
