+++ src/features/cookies/components/CookiesSettings.tsx
@@ -7,9 +7,9 @@ import { v4 as uuidv4 } from 'uuid'
 
 import { cookiesInfo } from 'features/cookies/components/cookiesInfo'
 import { CookieCategoriesEnum } from 'features/cookies/CookiesPolicy'
+import { useCookies } from 'features/cookies/helpers/useCookies'
+import { useCookiesChoiceByCategory } from 'features/cookies/helpers/useCookiesChoiceByCategory'
 import { CookiesSettingsProps } from 'features/cookies/pages/CookiesDetails'
-import { useCookies } from 'features/cookies/useCookies'
-import { useCookiesChoiceByCategory } from 'features/cookies/useCookiesChoiceByCategory'
 import { analytics } from 'libs/firebase/analytics'
 import { AccordionItem } from 'ui/components/AccordionItem'
 import FilterSwitch from 'ui/components/FilterSwitch'
