+++ src/features/cookies/components/useCookiesModalContent.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 
 import { CookiesConsentButtons } from 'features/cookies/components/CookiesConsentButtons'
 import { CookiesConsentExplanations } from 'features/cookies/components/CookiesConsentExplanations'
+import { CookiesChoiceByCategory } from 'features/cookies/helpers/useCookiesChoiceByCategory'
 import { CookiesDetails } from 'features/cookies/pages/CookiesDetails'
-import { CookiesChoiceByCategory } from 'features/cookies/useCookiesChoiceByCategory'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ModalLeftIconProps } from 'ui/components/modals/types'
 import { ArrowPrevious } from 'ui/svg/icons/ArrowPrevious'
