+++ src/features/cookies/pages/CookiesConsent.tsx
@@ -7,11 +7,11 @@ import {
   useCookiesModalContent,
 } from 'features/cookies/components/useCookiesModalContent'
 import { ALL_OPTIONAL_COOKIES, COOKIES_BY_CATEGORY } from 'features/cookies/CookiesPolicy'
-import { getCookiesChoiceFromCategories } from 'features/cookies/getCookiesChoiceFromCategories'
-import { startTracking } from 'features/cookies/startTracking/startTracking'
-import { startTrackingAcceptedCookies } from 'features/cookies/startTracking/startTrackingAcceptedCookies'
-import { useCookies } from 'features/cookies/useCookies'
-import { CookiesChoiceByCategory } from 'features/cookies/useCookiesChoiceByCategory'
+import { getCookiesChoiceFromCategories } from 'features/cookies/helpers/getCookiesChoiceFromCategories'
+import { startTracking } from 'features/cookies/helpers/startTracking'
+import { startTrackingAcceptedCookies } from 'features/cookies/helpers/startTrackingAcceptedCookies'
+import { useCookies } from 'features/cookies/helpers/useCookies'
+import { CookiesChoiceByCategory } from 'features/cookies/helpers/useCookiesChoiceByCategory'
 import { campaignTracker } from 'libs/campaign'
 import { analytics } from 'libs/firebase/analytics'
 import { requestIDFATrackingConsent } from 'libs/trackingConsent/useTrackingConsent'
